fix #560 many warnings fixed (Thanks to new PVS-Studio and cppcheck).

+++ modules/interpreter/src/cpp/ColonOperator.cpp
@@ -75,7 +75,7 @@ namespace Nelson {
 ArrayOf
 Evaluator::colonOperator(AbstractSyntaxTreePtr t)
 {
-    callstack.pushID(t->getContext());
+    callstack.pushID((size_t)t->getContext());
     ArrayOf retval = this->colonOperator(
         expression(t->down->down), expression(t->down->down->right), expression(t->down->right));
     callstack.popID();
@@ -85,7 +85,7 @@ Evaluator::colonOperator(AbstractSyntaxTreePtr t)
 ArrayOf
 Evaluator::colonUnitOperator(AbstractSyntaxTreePtr t)
 {
-    callstack.pushID(t->getContext());
+    callstack.pushID((size_t)t->getContext());
     ArrayOf retval = this->colonUnitOperator(expression(t->down), expression(t->down->right));
     callstack.popID();
     return retval;
