+++ modules/interpreter/src/cpp/MtimesOperator.cpp
@@ -32,7 +32,7 @@ namespace Nelson {
 ArrayOf
 Evaluator::mtimesOperator(AbstractSyntaxTreePtr t)
 {
-    callstack.pushID(t->getContext());
+    callstack.pushID((size_t)t->getContext());
     ArrayOf retval = this->mtimesOperator(expression(t->down), expression(t->down->right));
     callstack.popID();
     return retval;
